+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   issues during the restore and is now fixed.
 - restoring sessions multiple times messes up the whole environment - new panes
   are all around. This is now fixed - pane restorations are now idempotent.
+- if pane exists from before session restore - do not restore the process within
+  it. This makes the restoration process even more idempotent.
 
 ### v0.2.0, 2014-08-29
 - bugfix: with vim 'session' strategy, if the session file does not exist - make